Agent/requirements.txt
pn532pi
requests
paho-mqtt
python-plugwise